Move day toggle to own component and more schedule screen clean up
