+++ App/Containers/Styles/ScheduleScreenStyle.js
@@ -1,0 +1,34 @@
+import { StyleSheet } from 'react-native'
+import { ApplicationStyles, Metrics, Colors } from '../../Themes'
+
+export default StyleSheet.create({
+  ...ApplicationStyles.screen,
+  container: {
+    flex: 1,
+    backgroundColor: Colors.snow
+  },
+  row: {
+    flex: 1,
+    backgroundColor: Colors.snow,
+    marginVertical: Metrics.smallMargin
+  },
+  boldLabel: {
+    fontWeight: 'bold',
+    color: Colors.text
+  },
+  label: {
+    color: Colors.text
+  },
+  listContent: {
+    paddingTop: Metrics.baseMargin,
+    paddingBottom: Metrics.baseMargin * 8
+  },
+  timeline: {
+    width: 2,
+    backgroundColor: '#6E3C7B',
+    position: 'absolute',
+    top: 85,
+    bottom: 0,
+    right: 11
+  }
+})
